chore: move userDetailData
