+++ src/pages/settings/users/UserList.jsx
@@ -1,4 +1,4 @@
-import { useState, useMemo, useRef } from 'react'
+import { useMemo, useRef } from 'react'
 import { DataTable } from 'primereact/datatable'
 import { Column } from 'primereact/column'
 import { ContextMenu } from 'primereact/contextmenu'
@@ -7,46 +7,10 @@ import './users.sass'
 import useColumnResize from '/src/hooks/useColumnResize'
 import UserImage from './UserImage'
 
-const buildUserDetailData = (projectNames, roleNames, users, lastSelectedUser) => {
-  let roles = []
-  let roleSet = []
-  if (lastSelectedUser) {
-    if (!projectNames) roleSet = lastSelectedUser.defaultRoles || []
-    else {
-      const uroles = JSON.parse(lastSelectedUser.roles) || []
-      for (const projectName of projectNames || []) {
-        roleSet = [...roleSet, ...(uroles[projectName] || [])]
-      }
-    }
-  }
-
-  for (const roleName of roleNames)
-    roles.push({
-      name: roleName,
-      shouldSelect: roleSet.includes(roleName),
-    })
-
-  let userLevel = 'user'
-  if (lastSelectedUser?.isAdmin) userLevel = 'admin'
-  else if (lastSelectedUser?.isService) userLevel = 'service'
-  else if (lastSelectedUser?.isManager) userLevel = 'manager'
-
-  return {
-    users,
-    projectNames,
-    roles,
-    userLevel,
-    userActive: lastSelectedUser?.active,
-    isGuest: lastSelectedUser?.isGuest,
-  }
-}
-
 const UserList = ({
   selectedProjects,
   selectedUsers,
   setSelectedUsers,
-  setUserDetailData,
-  rolesList,
   userList,
   tableList,
   setShowRenameUser,
@@ -55,7 +19,6 @@ const UserList = ({
   isLoading,
   isLoadingRoles,
 }) => {
-  const [lastSelectedUser, setLastSelectedUser] = useState(null)
   const contextMenuRef = useRef(null)
 
   // COLUMN WIDTH
@@ -65,14 +28,8 @@ const UserList = ({
 
   const selection = useMemo(() => {
     let result = []
-    let lastUsr = null
     for (const user of userList) {
       if (selectedUsers.includes(user.name)) result.push(user)
-      if (user?.name === lastSelectedUser?.name) lastUsr = { ...user }
-    }
-    if (setUserDetailData) {
-      setLastSelectedUser(lastUsr)
-      setUserDetailData(buildUserDetailData(selectedProjects, rolesList, result, lastUsr))
     }
     return result
   }, [selectedUsers, selectedProjects])
@@ -122,12 +79,8 @@ const UserList = ({
               if (!selectedUsers.includes(e.value.name)) {
                 setSelectedUsers([...selection, e.value.name])
               }
-              setLastSelectedUser(e.data)
             }}
             selection={selection}
-            onRowClick={(e) => {
-              setLastSelectedUser(e.data)
-            }}
             columnResizeMode="expand"
             resizableColumns
             onColumnResizeEnd={setColumnWidths}
